requirements.txt
langchain
openai
tiktoken
langchain_community
pypdf
PyMuPDF
python-docx
python-dotenv
faiss-cpu